Update zigbee-sonoff-motion-sensor-expanded.groovy
Hi guys. Thanks a lot for your work. I want to suggest you a change of the format of battery attribute to an integer. Hubitat Dashboard with Smartly can not render a Battery tile properly when the value is in decimal format. The same may be relevant for other drivers then this one. Thanks. Michal